requirements.txt
aiogram==3.13.1
fastapi==0.115.0
uvicorn==0.30.6
aiosqlite==0.20.0
jinja2==3.1.4
python-multipart==0.0.9
starlette-session==0.4.3
pandas==2.2.2
openpyxl==3.1.5
python-dotenv